+++ CHANGES.md
@@ -1,3 +1,4 @@
 # CHANGES
 
+- Feature: Ported react-native-sound
 - Feature: Init